refactor: extract shared types
- Move all shared types to @copilot/shared/types/
- Update imports
