+++ package.json
@@ -60,5 +60,11 @@
 		"vue": "^3.5.22",
 		"webgl-plot": "^0.7.0",
 		"zod": "^4.3.5"
-	}
+	},
+	"private": true,
+	"workspaces": [
+		"src",
+		"web",
+		"packages/shared"
+	]
 }
