+++ packages/shared/package.json
@@ -1,0 +1,9 @@
+{
+    "name": "@copilot/shared",
+    "version": "1.0.0",
+    "main": "index.ts",
+    "types": "index.ts",
+    "dependencies": {
+        "@jlceda/pro-api-types": "^0.1.175"
+    }
+}
