+++ packages/shared/types/component.ts
