+++ packages/shared/types/eda.ts
@@ -1,7 +1,9 @@
-import { CircuitAssembly, ExplainCircuit } from "./circuit";
-import { SimulateResult } from "./spice";
-// @ts-ignore
+/* eslint-disable @typescript-eslint/ban-ts-comment */
+// @ts-expect-error
+// eslint-disable-next-line @typescript-eslint/no-unused-vars
 import type _ from '@jlceda/pro-api-types';
+import { CircuitAssembly, ExplainCircuit } from './circuit';
+import { SimulateResult } from './spice';
 
 declare global {
     interface EDA {
