+++ packages/shared/types/lcsc.ts
