+++ src/eda/types.ts
@@ -1,4 +1,4 @@
-import { CircuitAssembly } from "../types/circuit";
+import { CircuitAssembly } from "@copilot/shared/types/circuit";
 import { ComponentReplacer } from "./replacer";
 
 export const VCC_PORT_COMPONENT = {
