+++ web/src/main.ts
@@ -1,6 +1,7 @@
 import { createApp } from 'vue';
 import { createPinia } from 'pinia';
 import App from './App.vue';
+// @ts-ignore
 import 'katex/dist/katex.min.css';
 
 const app = createApp(App);
